chore: add rosetta functions to fetch and parse blocks and transactions
getRosettaBlockFromDataStore: used by /network/status and /block
Fetch a block by either its height or hash; if neither field is
provided, it returns the latest block.

getRosettaBlockTransactionsFromDataStore: used by /block

src/rosetta-helpers contains functions that operate on transaction
operations.

raw_result is in DbTx but not DbMempoolTx; add it to the latter to
make the interface consistent (needed for the Rosetta mempool APIs).