chore: make STACKS_NETWORK an ARG for the rosetta docker file
STACKS_NETWORK defaults to testnet, but can be set to mocknet,
dev, and testnet.  testnet and mocknet use the same Stacks-mocknet.toml
configuration file, and dev uses Stacks-dev.toml.  If it's set to something
else, it runs stacks-node krypton.

Tested with

    docker build -t stx-rosetta:stable -f stx-rosetta.Dockerfile .
    docker run --rm -p 3999:3999 --mount source=rosetta-data,target=/data \
	-e STACKS_NETWORK=dev stx-rosetta:stable
    docker run --rm -p 3999:3999 --mount source=rosetta-data,target=/data \
	-e STACKS_NETWORK=mainnet stx-rosetta:stable
    docker run --rm -p 3999:3999 --mount source=rosetta-data,target=/data \
	-e STACKS_NETWORK=testnet stx-rosetta:stable

and by calling

    curl -d '' -s http://localhost:3999/rosetta/v1/network/list

and

    docker exec -it the-container-id /bin/bash
    # echo $STACKS_NETWORK
    # ps auxww | grep stacks-node

This depends on a change in commit #db8e3ad44dbb45287a64d678f07de97644fc299c;
see line 11 in the stx-rosetta.Dockerfile to test without that change.